requirements.txt
streamlit==1.39.0
pydub==0.25.1
Pillow==10.4.0
moviepy==1.0.3
numpy==2.1.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-slugify==8.0.4
langid==1.1.6
imageio-ffmpeg==0.4.9
